move js code to a separate file
